i2c-stick-arduino/requirements.txt
pyyaml-include>=1.3.1
Jinja2>=3.1.2
PyYAML>=6.0.1
pyserial>=3.5
bumpver>=2023.1127
doit>=0.36.0
requests>=2.31.0
